% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split.R
\name{strsplit1}
\alias{strsplit1}
\title{Split a string and return the first element}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{A character string to be split.}

\item{split}{A character string giving the delimiter to split on.}
}
\value{
A character vector resulting from splitting \code{x}.
}
\description{
A small wrapper around \code{strsplit()} that returns only the first
element of the result instead of a list.
}
\examples{
strsplit1("a,b,c", split = ",")
strsplit1("hello world", split = " ")

}
